#56 新規投稿のコンポーネントを作成
